+++ test/e2e/page-objects/pages/settings/settings-page.ts
@@ -41,18 +41,18 @@ class SettingsPage {
     css: '.tab-bar__tab__content__title',
   };
 
+  private readonly backupAndSyncSettingsButton = {
+    text: 'Backup and sync',
+    css: '.tab-bar__tab__content__title',
+  };
+
   constructor(driver: Driver) {
     this.driver = driver;
   }
 
   async check_pageIsLoaded(): Promise<void> {
-    try {
-      await this.driver.waitForSelector(this.settingsPageTitle);
-    } catch (e) {
-      console.log('Timeout while waiting for Settings page to be loaded', e);
-      throw e;
-    }
-    console.log('Settings page is loaded');
+    console.log('Check settings page is loaded');
+    await this.driver.waitForSelector(this.settingsPageTitle);
   }
 
   async clickAdvancedTab(): Promise<void> {
@@ -94,8 +94,8 @@ class SettingsPage {
     await this.driver.clickElement(this.aboutViewButton);
   }
 
-  async goToDevelopOptionSettings(): Promise<void> {
-    console.log('Navigating to Develop options page');
+  async goToDeveloperOptions(): Promise<void> {
+    console.log('Navigating to Develop Options page');
     await this.driver.clickElement(this.developerOptionsButton);
   }
 
@@ -105,7 +105,7 @@ class SettingsPage {
   }
 
   async goToPrivacySettings(): Promise<void> {
-    console.log('Navigating to Privacy & Security Settings page');
+    console.log('Navigating to Privacy Settings page');
     await this.driver.clickElement(this.privacySettingsButton);
   }
 
@@ -117,6 +117,11 @@ class SettingsPage {
   async goToNotificationsSettings(): Promise<void> {
     console.log('Navigating to Notifications Settings page');
     await this.driver.clickElement(this.notificationsSettingsButton);
+  }
+
+  async goToBackupAndSyncSettings(): Promise<void> {
+    console.log('Navigating to Backup & Sync Settings page');
+    await this.driver.clickElement(this.backupAndSyncSettingsButton);
   }
 }